pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "wheel",
    "torch==2.1.0"
]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["pages*", "service*"]
exclude = ["db*", "models*", "huggingface*"]

[project]
name = "chatgpt-pdf"
version = "0.1.0"
description = "A comprehensive PDF processing and analysis application with OCR, AI, and document understanding capabilities"
requires-python = ">=3.11,<3.12"
dependencies = [
    "qdrant-client>=1.6.0,<1.10.0",
    "aiohttp==3.8.4",
    "aiosignal==1.3.1",
    "altair==5.0.1",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "async-timeout==4.0.2",
    "attrs==23.1.0",
    "beautifulsoup4==4.12.2",
    "blinker==1.6.2",
    "bs4==0.0.1",
    "cachetools==5.3.1",
    "certifi==2025.1.31",
    "cffi==1.15.1",
    "chardet==5.1.0",
    "charset-normalizer==3.2.0",
    "ci-info==0.3.0",
    "click==8.1.4",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "comtypes==1.2.0",
    "configobj==5.0.8",
    "configparser==6.0.0",
    "contourpy==1.3.3",
    "cryptography==41.0.2",
    "cycler==0.12.1",
    "dataclasses-json==0.5.9",
    "decorator==5.1.1",
    "depcheck==0.3.2",
    "distro==1.9.0",
    "et-xmlfile==1.1.0",
    "etelemetry==0.3.1",
    "exceptiongroup==1.2.2",
    "filelock==3.12.2",
    "filetype==1.2.0",
    "fitz==0.0.1.dev2",
    "flatbuffers==23.5.26",
    "fonttools==4.59.0",
    "frozenlist==1.4.0",
    "fsspec==2023.6.0",
    "future==0.18.3",
    "gitdb==4.0.10",
    "gitpython==3.1.32",
    "greenlet==2.0.2",
    "gtts==2.3.2",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "hub-sdk==0.0.20",
    "huggingface-hub==0.34.3",
    "humanfriendly==10.0",
    "idna==3.10",
    "imageio==2.37.0",
    "importlib-metadata==6.8.0",
    "isodate==0.6.1",
    "jinja2==3.1.2",
    "jiter==0.8.2",
    "joblib==1.3.1",
    "jsonschema==4.18.2",
    "jsonschema-specifications==2023.6.1",
    "kiwisolver==1.4.8",
    "loguru==0.7.3",
    "looseversion==1.3.0",
    "lxml==4.9.3",
    "markdown==3.4.3",
    "markdown-it-py==3.0.0",
    "markupsafe==2.1.3",
    "marshmallow==3.19.0",
    "marshmallow-enum==1.5.1",
    "matplotlib==3.10.5",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "msg-parser==1.2.0",
    "multidict==6.0.4",
    "mypy-extensions==1.0.0",
    "networkx==3.1",
    "nibabel==5.1.0",
    "nipype==1.8.6",
    "nltk==3.8.1",
    "numexpr==2.10.2",
    "numpy==1.26.4",
    "olefile==0.46",
    "onnx==1.18.0",
    "onnx-simplifier==0.4.36",
    "onnxruntime==1.19.2",
    "onnxsim==0.4.33",
    "openai==1.63.2",
    "openapi-schema-pydantic==1.2.4",
    "opencv-python==4.8.0.76",
    "openpyxl==3.1.2",
    "packaging>=16.8,<24",
    "pandas==2.0.3",
    "pathlib==1.0.1",
    "pdf2image==1.16.3",
    "pdfminer-six==20221105",
    "pdfplumber==0.9.0",
    "pillow==9.5.0",
    "protobuf>=3.20,<5",
    "prov==2.0.0",
    "psutil==7.0.0",
    "py-cpuinfo==9.0.0",
    "pyarrow==12.0.1",
    "pyclipper==1.3.0.post5",
    "pycparser==2.21",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydeck==0.8.1b0",
    "pydeps==1.12.20",
    "pydot==1.4.2",
    "pygments==2.15.1",
    "pympler==1.0.1",
    "pymupdf==1.23.5",
    "pymupdfb==1.23.5",
    "pypandoc==1.11",
    "pyparsing==3.1.1",
    "pypiwin32==223",
    "pyreadline3==3.4.1",
    "python-dateutil==2.8.2",
    "python-docx==0.8.11",
    "python-dotenv==1.0.0",
    "python-magic==0.4.27",
    "python-pptx==0.6.21",
    "pyttsx3==2.90",
    "pytz==2023.3",
    "pytz-deprecation-shim==0.1.0.post0",
    "pywin32==311",
    "pyxnat==1.6",
    "pyyaml==6.0",
    "rapid-orientation==0.0.8",
    "rapidocr-onnxruntime==1.4.4",
    "rdflib==7.0.0",
    "referencing==0.29.1",
    "regex==2023.6.3",
    "requests==2.31.0",
    "rich==13.4.2",
    "rpds-py==0.8.10",
    "safetensors==0.3.2",
    "scikit-learn==1.3.0",
    "scipy==1.11.2",
    "seaborn==0.13.2",
    "sentence-transformers==2.2.2",
    "sentencepiece==0.1.99",
    "shapely==2.0.2",
    "simplejson==3.19.2",
    "six==1.16.0",
    "smmap==5.0.0",
    "sniffio==1.3.1",
    "soupsieve==2.4.1",
    "sqlalchemy==2.0.18",
    "st-pages==0.4.3",
    "stdlib-list==0.11.1",
    "streamlit==1.25.0",
    "sympy==1.14.0",
    "tabulate==0.9.0",
    "tenacity==8.2.2",
    "thop==0.1.1.post2209072238",
    "threadpoolctl==3.2.0",
    "tiktoken==0.4.0",
    "tokenizers==0.13.3",
    "toml==0.10.2",
    "tomlkit==0.13.3",
    "toolz==0.12.0",
    "torch==2.1.0",
    "torchvision==0.16.0",
    "tornado==6.3.2",
    "tqdm==4.67.1",
    "traits==6.3.2",
    "transformers==4.32.0",
    "typing-extensions==4.12.2",
    "typing-inspect==0.9.0",
    "tzdata==2023.3",
    "tzlocal==4.3.1",
    "ultralytics==8.1.0",
    "unstructured==0.8.1",
    "urllib3==2.0.3",
    "validators==0.20.0",
    "wand==0.6.11",
    "watchdog==3.0.0",
    "win32-setctime==1.2.0",
    "xlrd==2.0.1",
    "xlsxwriter==3.1.2",
    "yarl==1.9.2",
    "yolo-minimal-inference==0.1.9",
    "zipp==3.16.1"
]

[tool.uv]
index = [
    { url = "https://pypi.tuna.tsinghua.edu.cn/simple", default = true }
]